+++ BaseReservation/BaseReservation.Application/Services/Interfaces/IServiceCanton.cs
@@ -9,7 +9,7 @@ public interface IServiceCanton
     /// </summary>
     /// <param name="provinceId">The ID of the province to filter cantons.</param>
     /// <returns>ICollection of ResponseCantonDto</returns>
-    Task<ICollection<ResponseCantonDto>> ListAllByProvinceAsync(byte provinceId);
+    Task<ICollection<ResponseCantonDto>> ListAllByProvinceAsync(long provinceId);
 
     /// <summary>
     /// Finds a canton by its unique ID.
@@ -17,5 +17,5 @@ public interface IServiceCanton
     /// <param name="id">The ID of the canton to retrieve. </param>
     /// <returns>ResponseCantonDto</returns>
     /// <exception cref="NotFoundException">Thrown when no canton is found with the specified ID</exception>
-    Task<ResponseCantonDto> FindByIdAsync(byte id);
+    Task<ResponseCantonDto> FindByIdAsync(long id);
 }
