+++ BaseReservation/BaseReservation.Application/Services/Interfaces/IServiceReservationDetial.cs
@@ -10,14 +10,14 @@ public interface IServiceReservationDetail
     /// </summary>
     /// <param name="reservationId">Branch id</param>
     /// <returns>ICollection of ResponseReservationDetailDto</returns>
-    Task<ICollection<ResponseReservationDetailDto>> ListAllByReservationAsync(int reservationId);
+    Task<ICollection<ResponseReservationDetailDto>> ListAllByReservationAsync(long reservationId);
 
     /// <summary>
     /// Get reservation detail with specific id
     /// </summary>
     /// <param name="id">Id to look for</param>
     /// <returns>ResponseReservationDetailDto</returns>
-    Task<ResponseReservationDetailDto?> FindByIdAsync(int id);
+    Task<ResponseReservationDetailDto?> FindByIdAsync(long id);
 
     /// <summary>
     /// Create reservation details
@@ -25,5 +25,5 @@ public interface IServiceReservationDetail
     /// <param name="reservationId">Branch id</param>
     /// <param name="reservationDetails">List of reservation detail model to be added</param>
     /// <returns>bool</returns>
-    Task<bool> CreateReservationDetailAsync(int reservationId, IEnumerable<RequestReservationDetailDto> reservationDetails);
+    Task<bool> CreateReservationDetailAsync(long reservationId, IEnumerable<RequestReservationDetailDto> reservationDetails);
 }
